+++ gcc/rust/Make-lang.in
@@ -131,6 +131,7 @@ GRS_OBJS = \
     rust/rust-default-resolver.o \
     rust/rust-toplevel-name-resolver-2.0.o \
     rust/rust-early-name-resolver-2.0.o \
+	rust/rust-finalize-imports-2.0.o \
     rust/rust-late-name-resolver-2.0.o \
 	rust/rust-immutable-name-resolution-context.o \
     rust/rust-early-name-resolver.o \
